Move KeyframeGroup back to being internal
