+++ Tests/SnapshotConfiguration.swift
@@ -38,11 +38,11 @@ extension SnapshotConfiguration {
       AnimationKeypath(keypath: "Checkmark Outlines.Group 1.Stroke 1.Color"): ColorValueProvider(.black),
       AnimationKeypath(keypath: "Checkmark Outlines 2.Group 1.Stroke 1.Color"): ColorValueProvider(.black),
       AnimationKeypath(keypath: "X Outlines.Group 1.Stroke 1.Color"): ColorValueProvider(.black),
-      AnimationKeypath(keypath: "Switch Outline Outlines.Fill 1.Color"): ColorValueProvider(.init(keyframes: [
+      AnimationKeypath(keypath: "Switch Outline Outlines.Fill 1.Color"): ColorValueProvider([
         Keyframe(value: Color.black, time: 0),
         Keyframe(value: Color(r: 0.76, g: 0.76, b: 0.76, a: 1), time: 75),
         Keyframe(value: Color.black, time: 150),
-      ])),
+      ]),
     ]),
 
     // Test cases for `AnimatedImageProvider`
